requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
SQLAlchemy==2.0.23
python-dotenv==1.0.0

# Production WSGI Server
gunicorn==21.2.0

# AI
# 向量检索（embedding 模块）
faiss-cpu>=1.7.0
numpy>=1.24.0
PyPDF2>=3.0.0
pycryptodome>=3.19.0
python-docx>=1.0.0
python-pptx>=0.6.0
# Excel 解析（.xlsx 用 openpyxl，.xls 用 xlrd）
openpyxl>=3.1.0
xlrd>=2.0.0
# 免费 OCR（知识库图片识别，需本机安装 Tesseract：Mac brew install tesseract tesseract-lang）
pytesseract>=0.3.10
Pillow>=10.0.0
faster-whisper>=1.0.0
flask-sock>=0.7.0
edge-tts>=6.1.0
# 文生图（Stable Diffusion XL，diffusers）
torch>=2.0
torchvision
diffusers>=0.30.0
transformers>=4.46.0
accelerate
qwen-vl-utils
av
dashscope>=1.25.0
qwen-agent>=0.0.33
# LangChain SQL Agent（Text2SQL，无需手写提示词）
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
# LangGraph 图工作流（循环、并行、条件路由、人机协作等）
langgraph>=0.2.0
python-dateutil==2.8.2
mcp>=1.26.0
# 阿里云 DNS SDK（PyPI 包名与 import 名不同）
aliyun-python-sdk-core>=1.2.11
aliyun-python-sdk-alidns>=3.0.0